+++ AdapterUnitTests/InMobiAdapterTests/AUTInMobiInterstitialAdTest.m
@@ -55,7 +55,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
 
@@ -96,7 +97,8 @@ GADMediationInterstitialAdConfiguration
                            extras:extras];
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   NSData *bidResponseData = [bidResponse dataUsingEncoding:NSUTF8StringEncoding];
   OCMStub([_interstialMock load:bidResponseData]).andDo(^(NSInvocation *invocation) {
@@ -154,7 +156,8 @@ GADMediationInterstitialAdConfiguration
                            extras:extras];
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   NSData *bidResponseData = [bidResponse dataUsingEncoding:NSUTF8StringEncoding];
   OCMStub([_interstialMock load:bidResponseData]).andDo(^(NSInvocation *invocation) {
@@ -221,7 +224,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   // Mock unsuccessful ad loading.
   IMRequestStatus *expectedError = OCMClassMock([IMRequestStatus class]);
@@ -278,7 +282,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =
@@ -308,7 +313,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =
@@ -338,7 +344,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =
@@ -362,7 +369,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =
@@ -383,7 +391,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =
@@ -403,7 +412,8 @@ GADMediationInterstitialAdConfiguration
       AUTGADMediationInterstitialAdConfigurationForInMobi();
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment, nil,
+      GADAgeRestrictedTreatmentUnspecified);
   [self mockIMInterstitialWithRequestParameters:requestParameters];
   [self mockSuccessfulAdLoading];
   id<GADMediationInterstitialAdEventDelegate> interstitalAdEventDelegate =