+++ AdapterUnitTests/InMobiAdapterTests/AUTInMobiUtilsTest.m
@@ -43,6 +43,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
 - (void)tearDown {
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = nil;
   GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = nil;
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentUnspecified;
   [NSUserDefaults.standardUserDefaults removeObjectForKey:@"IABUSPrivacy_String"];
   [super tearDown];
 }
@@ -293,6 +295,34 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   OCMVerifyAll(IMSDKMock);
 }
 
+- (void)testTargetingFromAdConfiguration_WhenAgeRestrictedTreatmentIsChild {
+  GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment =
+      GADAgeRestrictedTreatmentChild;
+
+  GADInMobiExtras *extras = AUTGADInMobiExtras();
+  GADMediationAdConfiguration *configuration =
+      [[GADMediationAdConfiguration alloc] initWithAdConfiguration:nil
+                                                         targeting:nil
+                                                       credentials:OCMOCK_ANY
+                                                            extras:extras];
+
+  id IMSDKMock = OCMClassMock([IMSdk class]);
+  OCMExpect([IMSDKMock setLocation:extras.location]);
+  OCMExpect([IMSDKMock setPostalCode:extras.postalCode]);
+  OCMExpect([IMSDKMock setAreaCode:extras.areaCode]);
+  OCMExpect([IMSDKMock setInterests:extras.interests]);
+  OCMExpect([IMSDKMock setAge:extras.age]);
+  OCMExpect([IMSDKMock setYearOfBirth:extras.yearOfBirth]);
+  OCMExpect([IMSDKMock setLanguage:extras.language]);
+  OCMExpect([IMSDKMock setEducation:extras.educationType]);
+  OCMExpect([IMSDKMock setAgeGroup:extras.ageGroup]);
+  OCMExpect([IMSDKMock setIsAgeRestricted:YES]);
+
+  GADMAdapterInMobiSetTargetingFromAdConfiguration(configuration);
+
+  OCMVerifyAll(IMSDKMock);
+}
+
 - (void)testTargetingFromAdConfiguration_WhenTagForChildIsTrueAndUnderAgeIsFalse {
   GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment = @YES;
   GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent = @NO;
@@ -388,7 +418,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeWaterfall,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
   NSString *versionString =
       [NSString stringWithFormat:@"afma-sdk-i-v%ld.%ld.%ld",
                                  GADMobileAds.sharedInstance.versionNumber.majorVersion,
@@ -412,7 +443,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
   NSString *versionString =
       [NSString stringWithFormat:@"afma-sdk-i-v%ld.%ld.%ld",
                                  GADMobileAds.sharedInstance.versionNumber.majorVersion,
@@ -438,7 +470,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"1");
 }
@@ -456,7 +489,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"0");
 }
@@ -474,7 +508,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"1");
 }
@@ -492,7 +527,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"0");
 }
@@ -511,7 +547,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"1");
 }
@@ -530,7 +567,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"1");
 }
@@ -549,7 +587,8 @@ GADInMobiExtras *_Nonnull AUTGADInMobiExtras() {
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       configuration.extras, GADMAdapterInMobiRequestParametersMediationTypeRTB,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
 
   XCTAssertEqualObjects([requestParameters valueForKey:@"coppa"], @"1");
 }