+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiInterstitialAd.m
@@ -121,7 +121,8 @@
   NSDictionary<NSString *, id> *requestParameters = GADMAdapterInMobiRequestParameters(
       extras, mediationType,
       GADMobileAds.sharedInstance.requestConfiguration.tagForChildDirectedTreatment,
-      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent);
+      GADMobileAds.sharedInstance.requestConfiguration.tagForUnderAgeOfConsent,
+      GADMobileAds.sharedInstance.requestConfiguration.ageRestrictedTreatment);
   [_interstitialAd setExtras:requestParameters];
 
   if (bidResponseData) {