+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobiUtils.h
@@ -62,7 +62,8 @@ void GADMAdapterInMobiSetTargetingFromAdConfiguration(
 NSDictionary<NSString *, id> *_Nonnull GADMAdapterInMobiRequestParameters(
     GADInMobiExtras *_Nullable extras,
     GADMAdapterInMobiRequestParametersMediationType _Nonnull mediationType,
-    NSNumber *_Nullable childDirectedTreatment, NSNumber *_Nullable underAgeOfConsent);
+    NSNumber *_Nullable childDirectedTreatment, NSNumber *_Nullable underAgeOfConsent,
+    GADAgeRestrictedTreatment *ageRestrictedTreatment);
 
 /// Logs with GADMediationAdapterInMobi Prefix.
 void GADMAdapterInMobiLog(NSString *_Nonnull format, ...);